Rename helper modules.
