+++ sandpiper/tests/conftest.py
@@ -8,8 +8,8 @@ import discord.ext.commands as commands
 import pytest
 import pytz
 
-from .helpers._discord_helpers import *
-from .helpers._helpers import patch_all_symbol_imports
+from .helpers.discord import *
+from .helpers.misc import patch_all_symbol_imports
 from sandpiper.user_data import DatabaseSQLite
