+++ sandpiper/tests/test_database.py
@@ -4,7 +4,7 @@ from typing import Optional
 import pytest
 import pytz
 
-from .helpers._helpers import *
+from .helpers.misc import *
 from sandpiper.common.time import TimezoneType
 from sandpiper.user_data import *
